requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.1.0
gunicorn==20.1.0
platformdirs==2.4.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.8.1
whitenoise==5.3.0
